Refactor prefix management: create fetchPrefixes API, implement PrefixView component, and update registry imports
