+++ src/endpoints/poisontype/PoisonTypeView.tsx
