+++ src/types/prefix.ts
@@ -1,9 +1,3 @@
-
-
 export interface PrefixesPayload {
   prefixes: string[];
 }
-
-
-
-
